feat(uc_protocol): relay the frame header — client_id/seq u32, time_ns u64; FRAME_TYPE_TIMER + TimerBody codec (tests: field_offsets_are_the_relaid_layout, header_roundtrip_except_length_pins_the_bytes, timer_body_roundtrip_and_short_input_is_none)

+++ uc_log/examples/m1_gate.rs
@@ -95,7 +95,7 @@ fn main() {
     let mut samples: Vec<(f64, u64)> = Vec::new();
     while Instant::now() < deadline {
         for _ in 0..1024 {
-            match appender.append(1, appended, &payload) {
+            match appender.append(1, appended as u32, &payload) {
                 Ok(_) => appended += 1,
                 Err(AppendError::WouldOverrun) => {
                     overruns += 1;
